FINAL_PIPELINE/requirements.txt
google-generativeai
typing-extensions
pathlib2
vertex-ai
